Do not download files that already exists
As the program could be invoked many times in example time range and the
new wallpaper might not arrived at that time, it is possible that we
would fetch the same image many times.  Which is unnecessary bandwidth
and time lost.

By combining subreddit name and submission id into the file name of
saved wallpaper, we can verify on the next fetch if the wallpaper was
already saved.